requirements.txt
streamlit
pandas
supabase
openpyxl
google-generativeai